requirement.txt
asgiref==3.8.1
certifi==2024.8.30
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.3
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
gunicorn==23.0.0
mysql-connector-python==9.1.0
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.1
whitenoise==6.8.2
